Init File Service

+++ backend/app-spring/src/main/java/aaagt/cloudservice/file/dto/ListResponseFileItemDto.java
@@ -2,5 +2,5 @@ package aaagt.cloudservice.file.dto;
 
 public record ListResponseFileItemDto(
         String filename,
-        Integer size
+        Long size
 ) {}
